xyyxf poly generation had some issues: made scratch algebraic coefficients signed 64 bit integers so coefficients overflow less often and fixed issue with treating composite bases.  started making things compatible with visual studio builds again, but msieve's explicit use of pthreads is hosing it for now.

+++ include/nfs.h
@@ -137,7 +137,7 @@ typedef struct
 
 
 	mpz_polys_t* poly;
-	int c[MAX_POLY_DEGREE + 1]; // scratch space -- converted to mpz_poly_t
+	int64 c[MAX_POLY_DEGREE + 1]; // scratch space -- converted to mpz_poly_t
 				    // in check_poly()
 
 	// other useful parameters
